got image to render on page upon adding
