+++ client/.eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/index.js":"1","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/App.js":"2","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/index.js":"3","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/searchBar.js":"4","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/navBar.js":"5","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/allImages.js":"6","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/singleImg.js":"7","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/addImageForm.js":"8"},{"size":198,"mtime":1610750072425,"results":"9","hashOfConfig":"10"},{"size":1336,"mtime":1610923355761,"results":"11","hashOfConfig":"10"},{"size":138,"mtime":1610736506479,"results":"12","hashOfConfig":"10"},{"size":576,"mtime":1610922037728,"results":"13","hashOfConfig":"10"},{"size":1243,"mtime":1610921944850,"results":"14","hashOfConfig":"10"},{"size":319,"mtime":1610921145648,"results":"15","hashOfConfig":"10"},{"size":683,"mtime":1610920289237,"results":"16","hashOfConfig":"10"},{"size":1645,"mtime":1610917953710,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"edcrm",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/index.js",[],["35","36"],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/App.js",[],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/index.js",[],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/searchBar.js",[],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/navBar.js",[],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/allImages.js",[],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/singleImg.js",[],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/addImageForm.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/index.js":"1","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/App.js":"2","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/index.js":"3","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/searchBar.js":"4","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/navBar.js":"5","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/allImages.js":"6","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/singleImg.js":"7","/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/addImageForm.js":"8"},{"size":198,"mtime":1610750072425,"results":"9","hashOfConfig":"10"},{"size":1552,"mtime":1610924573959,"results":"11","hashOfConfig":"10"},{"size":138,"mtime":1610736506479,"results":"12","hashOfConfig":"10"},{"size":576,"mtime":1610922037728,"results":"13","hashOfConfig":"10"},{"size":1278,"mtime":1610923956787,"results":"14","hashOfConfig":"10"},{"size":319,"mtime":1610921145648,"results":"15","hashOfConfig":"10"},{"size":683,"mtime":1610920289237,"results":"16","hashOfConfig":"10"},{"size":1795,"mtime":1610924318529,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"edcrm",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/index.js",[],["35","36"],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/App.js",[],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/index.js",[],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/searchBar.js",[],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/navBar.js",[],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/allImages.js",[],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/singleImg.js",[],"/Users/khuong/Desktop/Projects/projects/shopify/image-repo/client/src/components/addImageForm.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
