+++ server/api/index.js
@@ -2,6 +2,9 @@ const router = require('express').Router()
 const {Image} = require('../db/models')
 module.exports = router
 
+
+router.use('/upload', require('./upload'))
+
 router.get('/', async (req, res, next) => {
   try {
     const images = await Image.findAll()
@@ -10,6 +13,7 @@ router.get('/', async (req, res, next) => {
     next(error)
   }
 });
+
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
